Moved RepiEquivalent in hazardlib.valid
